requirements.txt
# Core dependencies for RAG chatbot
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
openai==1.7.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
